improvement: S3UTILS-32 option ONLY_DELETED to cleanupNoncurrentVersions
Add an environment option ONLY_DELETED to the
cleanupNoncurrentVersions script, so that it only deletes noncurrent
versions if the object's current version is a delete marker. It also
then deletes the delete marker as usual.

Also updated the README.md to reflect the new option and fixed a few
missing updates.