Switch to alpine image
This is needed to address CVEs in glibc package, which are not fixed in
debian buster; as well as in python, which are not fixed in debian
bullseye.

Reduce size of image, as well as the (most likely unused) dependencies
we silently get from base node.js image.

Issue: S3UTILS-69