bugfix: S3UTILS-85 enforce version ID suffix
Always output versioned keys with their version suffix, even if we are
processing the master version. The change occurs in all places that
are used for divergence detection:

- in verifyBucketSproxydKeys to generate the digests database

- in the BucketStream and DBListStream helper classes used for the
  diff algorithm between a bucket listing and raw databases

This is useful to have a cleaner diff output and avoids some
shenanigans later during the repair process.

Update the procedure in README.md to point to the correct s3utils
image version.