Fix handling of null-version-id objects
When an object does not have a null-version-id, the "fix" was not
correct, as we did not set all fields (`isNull` and `nullVersionId`).
This is detailed in [1].

Upgrading is not actually needed to set the REPLICATION status, so the
default is now to skip this format change, and just update the metadata
in the object.

Alternatively, it is also possible to set GENERATE_INTERNAL_VERSION_ID
env variable so that these objects actually get migrated to the new
"versionned" format, with an allocated internal version id, but while
keeping the external "null" version id.

[1] https://github.com/scality/cloudserver/blob/development/8.5/docs/ARCHITECTURE.rst#null-version-management

Issue: S3UTILS-77