improvement: S3UTILS-90 update followerDiff with oplog filtering
Update the CompareRaftMembers/followerDiff script to accept a new
environment variable EXCLUDE_FROM_CSEQS: if passed with the current
cseq mapping of the raft sessions being scanned, it filters the diff
entries matching a key present in the corresponding bucket's raft
session oplog more recently than the given cseq.

The goal is to filter false positives due to the leader being ahead of
the follower being scanned (because the follower repds are stopped,
they cannot receive the newer updates).