impr: S3UTILS-138 detect duplicate versionId field
In master keys, detect when the metadata blob has more than one
`versionId` field, with a regular expression match. It is not a
completely correct method but it should work in practice, and
JSON.parse cannot be used for this purpose.

The output log line will be able to be parsed by another tool to
repair the affected objects.