Fix handling of restored objects
The code was trying to find out the "cold" location of restored objects
by looking at the `value.location` field which is neither projected nor
contains that information.

In earler Zenko releases (up to 2.6.34), the information was simply
not available. In 2.6.35 however, code was updated to work similar to
AWS (reporting the cold location via the storage class), and the name of
the location was thus added to the `x-amz-storage-class` field.

Issye: S3UTILS-141