Ignore blank PHD marker
In some case, PHD are re-created form scratch and do not hold any
information (when deleting a delete marker which is the latest version).
This should be very transient (they should be cleaned after 5 seconds),
but should not create any error and we can just skip the entry.

We should also skip any error from `_processEntryData()`, as they should
not really happen (only corrupted DB or coding issue), and should thus
not prevent the process from completing.

Issue: S3UTILS-142