S3UTILS-163: adapt inflight support
We want to compute the diff between the inflight for each bucket
at the beginning of its scan with the inflight at the end of the
count-items job. this ensure that we do not lose any information
between the start of a scan and the end of count-items, that would
cause quotas to be exceeded as inflights are lost.

The limitation of this implementation comes from the fact that
inflights are not created in buckets when there is only quotas on
Accounts. This is something we will improve on Scuba:Cloudserver
side; so the introduced implementation will be future-proof.