Avoid duplicate countitems metrics
Metrics from both workers & managers are instantiated in both kind of
processes, leading to duplicates when merging (naively) the metrics.

Improving merge is not trivial, for now we simply are more careful about
instantiating the metrics: `bucketProcessingDuration` and
`consolidationDuration` are only instantiated on master process. Since
they have labels, `bucketsCount` and `objectsCount` are not affected, so
we can keep them as is for now (since they are used directly from
S3UtilsMongoClient, and would need a significant refactor to make it
work both in prod and tests).

Issue: S3UTILS-175