Fix CrrExistingObjects skipping versions when listing
The listing function from metadata expects versionIdMarker to equal the non encoded
versionId of the next version to list. However, currently we give it the encoded
versionId, which causes the listing algo to skip some versions, as the versionId
is used to skip all previous versions until versionIdMarker is reached. Using the
encoded versionId in the check results in unpredictable behaviour that can skip
multiple versions from the listing.

Issue: S3UTILS-185