crrExistingObjects: support only triggering current versions
This is needed when using multiple-backend replication, as
triggering all versions may result in an ordering missmatch
between versions on both sites. This is due to the parallel
execution of PUTs to the destination, there is no guarantee
the order of versions in source will be respected, the latest
version PUT in destination becomes the master.

Issue: S3UTILS-183